// frontend/src/openapi/models/connection.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ReadersLounge API
 * ReadersLounge API
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface Connection
 */
export interface Connection {
  /**
   *
   * @type {number}
   * @memberof Connection
   */
  connection_id: number;
  /**
   *
   * @type {number}
   * @memberof Connection
   */
  target_user_id: number;
  /**
   *
   * @type {string}
   * @memberof Connection
   */
  target_user_name: string;
  /**
   *
   * @type {string}
   * @memberof Connection
   */
  target_user_profile_image: string;
}